Add: readd the directory: "Hack_Fix", Into that folder you find NOT Official MoDB Patch, that solve current Abuse Problem because lack of support from the core or db.
So all patchs in there, are OPTIONAL, But normally Suggested by they poster as good course of action to keep running without abuse problem.

Simply Update you DB with the "DEL-xxxxxxxx.sql" patch file to get rid of that abuse problem.

Do "ADD-xxxxxxxx.sql" to Readd The Deleted Data.

When abuse probleme is Solved, we will delete: "DEL-xxxxxx.sql" from SVN, at this time just run the "ADD", If you got error on "ADD", that mean it ok Don't need to run it ;)